+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fixed assertion error on closing or invalidating a link with messages forwarded from a virtual link.
 - Fixed host hanging if stopped while an instance is in the process of creating a save during startup.
 - Fixed `instance.auto_start` option.
+- Fixed host log spammed with errors when plugins broadcast messages to instances with the plugin disabled.
 
 
 ## Version 2.0.0-alpha.14